remove explicit getty on ttyS2
systemd should spawn the getty if console=ttyS2 is present on the
kernel command line. This does seem to work so there isn't any need
to explicitely enable the service.

It does skip the 115200 baud option for some reason. Try lower baud
rates if 115200 doesn't work, e.g. 57600 or 38400. This shouldn't
be related to the settings we are removing in this commit.